some refactor, added tons of comments to look at coming refactor
