+++ src/main/java/com/sellmultiplier/utils/Multiplier.java
@@ -1,4 +1,4 @@
-package com.sellmultiplier;
+package com.sellmultiplier.utils;
 
 import java.math.BigDecimal;
